move calculators to state folders
