+++ programs/programs/co/__init__.py
@@ -1,0 +1,56 @@
+from .rtdlive.calculator import RtdLive
+from .child_care_assistance.calculator import ChildCareAssistance
+from .mydenver.calculator import MyDenver
+from .cash_back.calculator import CashBack
+from .energy_assistance.calculator import EnergyAssistance
+from .energy_resource_center.calculator import EnergyResourceCenter
+from .omnisalud.calculator import OmniSalud
+from .dental_health_care_seniors.calculator import DentalHealthCareSeniors
+from .reproductive_health_care.calculator import ReproductiveHealthCare
+from .connect_for_health.calculator import ConnectForHealth
+from .medicaid.family_planning_services.calculator import FamilyPlanningServices
+from .denver_preschool_program.calculator import DenverPreschoolProgram
+from .every_day_eats.calculator import EveryDayEats
+from .property_credit_rebate.calculator import PropertyCreditRebate
+from .universal_preschool.calculator import UniversalPreschool
+from .my_spark.calculator import MySpark
+from .low_wage_covid_relief.calculator import LowWageCovidRelief
+from .medicaid.child_with_disability.calculator import MedicaidChildWithDisability
+from .medicaid.adult_with_disability.calculator import MedicaidAdultWithDisability
+from .medicaid.emergency.calculator import EmergencyMedicaid
+from .basic_cash_assistance.calculator import BasicCashAssistance
+from .weatherization_assistance.calculator import WeatherizationAssistance
+from .tabor.calculator import Tabor
+from .trua.calculator import Trua
+from .utility_bill_pay.calculator import UtilityBillPay
+from ..calc import ProgramCalculator
+from .rental_assistance_grant.calculator import RentalAssistanceGrant
+
+co_calculators: dict[str, type[ProgramCalculator]] = {
+    'rtdlive': RtdLive,
+    'cccap': ChildCareAssistance,
+    'mydenver': MyDenver,
+    'cocb': CashBack,
+    'leap': EnergyAssistance,
+    'erc': EnergyResourceCenter,
+    'omnisalud': OmniSalud,
+    'cdhcs': DentalHealthCareSeniors,
+    'rhc': ReproductiveHealthCare,
+    'cfhc': ConnectForHealth,
+    'fps': FamilyPlanningServices,
+    'dpp': DenverPreschoolProgram,
+    'ede': EveryDayEats,
+    'cpcr': PropertyCreditRebate,
+    'upk': UniversalPreschool,
+    'myspark': MySpark,
+    'lwcr': LowWageCovidRelief,
+    'cwd_medicaid': MedicaidChildWithDisability,
+    'awd_medicaid': MedicaidAdultWithDisability,
+    'emergency_medicaid': EmergencyMedicaid,
+    'bca': BasicCashAssistance,
+    'cowap': WeatherizationAssistance,
+    'tabor': Tabor,
+    'trua': Trua,
+    'ubp': UtilityBillPay,
+    'rag': RentalAssistanceGrant,
+}
